Move test fixture value into pcode-ops
