+++ crates/pcode-ops/src/lib.rs
@@ -2,4 +2,7 @@ pub mod convert;
 pub mod validate;
 
 mod ops;
+mod pcode128;
+
 pub use ops::*;
+pub use pcode128::*;
